REL-4282 - ITC calculation of exposure time & number of exposures to achieve S/N

+++ bundle/edu.gemini.itc/src/main/java/edu/gemini/itc/operation/ImagingPointS2NMethodBCalculation.java
@@ -40,7 +40,7 @@ public final class ImagingPointS2NMethodBCalculation extends ImagingS2NCalculati
 
     }
 
-    @Override public double numberSourceExposures() {
+    @Override public int numberSourceExposures() {
         return int_req_source_exposures;
     }
